fix: Resolve additional MSVC C4267 warnings in Windows builds
Fixed size_t to int conversion warnings in multiple files:

**footer_renderer.cpp (2 locations):**
- Line 93: Cast label.length() for std::min parameter
- Line 108: Cast loop index for write_char column parameter

**point_editor.cpp (2 locations):**
- Line 72: Cast points.size() return value to int
- Line 139: Cast points.size() return value to int

**table_view.cpp (1 location):**
- Line 141: Cast get_visible_rows().size() return value to int

**table_selection_menu.cpp (16 locations):**
- Lines 22, 30, 44, 63, 163: Cast string length for column calculations
- Lines 26, 33, 48, 65, 116, 127, 154, 165: Cast loop indices for write_char
- Lines 131, 190, 195: Cast tables.size() to int for action indexing

**table_creation_dialog.cpp (2 locations):**
- Line 158: Cast current_field_ for cursor row calculation
- Line 159: Cast cursor_pos_ for cursor column calculation

All changes use explicit static_cast<int>() for MSVC /W4 /WX compliance.

+++ src/footer_renderer.cpp
@@ -90,7 +90,7 @@ void FooterRenderer::render(Terminal& terminal, double cursor_x, double cursor_y
                 footer_str.resize(cols, ' ');
             }
             int start = std::max(0, cols - static_cast<int>(label.length()));
-            footer_str.replace(start, std::min<int>(label.length(), cols - start), label);
+            footer_str.replace(start, std::min<int>(static_cast<int>(label.length()), cols - start), label);
         }
     };
 
@@ -105,7 +105,7 @@ void FooterRenderer::render(Terminal& terminal, double cursor_x, double cursor_y
     }
 
     for (size_t i = 0; i < footer_str.length(); ++i) {
-        terminal.write_char(footer_row, i, footer_str[i]);
+        terminal.write_char(footer_row, static_cast<int>(i), footer_str[i]);
     }
 }
